pyproject.toml
[project]
name = "wedding-os-backend"
version = "0.1.0"
description = "Wedding OS Backend - FastAPI 기반 백엔드 서버"
readme = "../README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.121.3",
    "uvicorn[standard]>=0.38.0",
    "sqlalchemy>=2.0.44",
    "pymysql>=1.1.2",
    "aiomysql>=0.3.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.2.1",
    "bcrypt>=5.0.0",
    "httpx>=0.28.1",
    "chromadb>=1.3.5",
    "langchain-core>=1.1.0",
    "langchain-chroma>=1.0.0",
    "qrcode>=8.2",
    "reportlab>=4.4.5",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true




